heroku app configuration and backgrounds
First, this commit adds Procfiles to both the frontend and the backend, deployed as separate webapps, for Heroku to use. It also cleans up a bit of the yarn/npm conflicts that prevented these from being deployed on the web.

Second, this commit adds a background and various text upgrades to the site itself. This helps it to look more pretty and more like a 'real website'